requirements-fastapi.txt
fastapi==0.115.12
uvicorn==0.34.2
numpy==2.2.5
faiss-cpu==1.13.2
sentence-transformers==4.1.0
huggingface_hub==0.36.2
groq==0.24.0
python-dotenv==1.1.0
